CGTeleport: is now CGMonolith / CGSubterraneanGate / CGWhirlpool
This change is needed for the future pathfinding improvements, but for now just split functionality we currently have into several classes.
Though in future CGSubterraneanGate won't have own onHeroVisit as it's may be represented as two-way monolith.

Other target of this refactoring is make it possible to have subtype of Monolith that include both unidirectional and bidirectional objects.