Move generic CGMonolith base CGTeleport class
Originally I didn't want to do that, but now I see that it's better to be sure that code called from CGMonolith actually handle it's specific and nothing else.
With new channels feature abd related changes previous naming it's even more pointless.
Keep in mind that code not going to compile in this commit.