VCAI: always use addVisitableObj to insert anything into visitableObjs
This is needed because we want to always add any new teleport channels to knownTeleportChannels before object is used for any purpose.
In past there was crash in Explore::getAllPossibleSubgoals because teleport was used before it's channel added to known list.
This should help to avoid crash in Explore::getAllPossibleSubgoals